added contact and books endpoints
